requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
djangorestframework==3.14.0
fonttools==4.39.4
freetype-py==2.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
psycopg==3.1.9
psycopg2==2.9.6
pycairo==1.23.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.0
rlPyCairo==0.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
